Create static version of richtext
